Burmese-object-language lemmas from Burmese-meta-language Wiktionary.
The words in this list occur within the titles of non-redirect pages
in the main namespace of the Burmese Wikitionary (dump form 2016-09-20).
We only retained titles where the page text gives some indication that the
title can be a Burmese word (so as to exclude words from other languages
written in the Myanmar script).

The lemmas were slightly post-processed to eliminate common mistakes
in the representation of Burmese such as misspellings, repeated diacritics,
non-canonical ordering, etc.