Further encapsulate matrix handling
WORLD assumes an array representation of vectors and matrices, but for
many reasons (including memory management) we prefer representations
built on top of std::vector. This commit adds a small helper class
that makes it possible to access the same data as a vector of vectors or
as a two-dimensional array.