Refactor aperiodicity computation
Refactor Bap() to use MatrixHelper along the same lines as
Spectrogram(). The automatic memory management this entails also fixes
the memory leak due to an explicit allocation with 'new' which had not
been paired with a corresponding 'delete'.

Fix(?) bap dimension, at least to the point where it is internally
consistent. This at least uses a consistent (if possibly incorrect)
inner bap dimension, where an inconsistent (and definitely incorrect)
num_frames_ had been present before. It also eliminates an out-of-bounds
array access in the highest per-frame aperiodicity index.