Travis CI: Factor out script, use only 2 Pythons
Move Travis script logic into a dedicated shell script.

Reduce the number of Python versions in build matrix. There was no
reason to go for the Full Monty of all Pythons, as this greatly
increased the total runtime and did not provide useful information.

Settled on Python 2.7 and 3.3, as 2.7 is the final minor version of
Python 2 and 3.3 is the minimally required minor version of Python 3.