Further tweak SymbolTable generation script
Include Sinhala alternative names. Sinhala short names are then derived
in a standard fashion from alternative (Sanskritic) names. The official
(Sinhalese) names do not match the conventions for the other Indic
scripts.

Do not mangle names of vocalic consonants. We now write "vocalic_r"
instead of "ri" to make it easier to match the short name with the
Unicode name. The previous short name "ri" was intended as a quick
transliteration, but as we are replacing this with a more systematic
transliteration approach, this use case is no longer relevant.

Always include ZWNJ and ZWJ in every script symbol table.