Travis CI: Downgrade Bazel to 0.22.0
Recent Bazel releases ship with embedded OpenJDK.
Bazel release 0.22.0 is the last release to use OpenJDK 9.
Subsequent Bazel releases ship with OpenJDK 11.

Android SDK manager is not compatible with JDK 11, since it relies
on the package javax.xml.bind.annotation, formerly provided by
Java EE and still accessible via modules in JDK 9. This no
longer works in JDK 11, which removes Java EE modules.

A short-term workaround would be to side-load JAXB, e.g. fetching
it from a Maven coordinate and setting up the appropriate CLASSPATH.

Instead we'll aim for the longer-term fix: wait for the Android SDK
tools to catch up to Java 11 (or ship with the required dependencies).